Rename Verifiable to ExpectedFixity; move to the bag verifier
