+++ bag_verifier/src/main/scala/uk/ac/wellcome/platform/archive/bagverifier/fixity/ExpectedFixity.scala
@@ -1,0 +1,13 @@
+package uk.ac.wellcome.platform.archive.bagverifier.fixity
+
+import uk.ac.wellcome.platform.archive.common.verify.VerifiableLocation
+
+/** Given some Container of files, get the expected fixity information (size/checksum)
+  * for every file in the container.
+  *
+  */
+trait ExpectedFixity[Container] {
+  def create(container: Container): Either[CannotCreateExpectedFixity, Seq[VerifiableLocation]]
+}
+
+case class CannotCreateExpectedFixity(msg: String) extends Throwable(msg)
